Chrome Store ● Pending Review
Permissions restricted to minimum requirements:
● "activeTab" ● bg_content_scripts_loaded_update  ● get current tab url
" ● To get the current tab URL in order to activate the content-script on a per-page user choice.

● "scripting" ● bg_content_scripts_loaded_update  ● get content-script state and reload command
" ● To activate or deactivate the content script when the user clicks on the Popup start/stop button.
" ● For pages that were already loaded when the extension was last activated, a location reload will inject the content script.

● "storage"   ● bg_tabs_onActivated               ● get current page working attributes
" ● To get the current page user's options for that page.

● "tabs"      ● bg_tabs_get_url                   ● get a tab url
" ● To get the tab's url as a key for user's per-page stored options.

● "Host permissions"
" ● content_scripts matches:["<all_urls>"] looks like being a mandatory requirement in manfest.json V3.